--- app/components/home-component.tsx
@@ -1,10 +1,0 @@
-import { ContentContainerComponent } from "@justaddjam/strawberry";
-import * as React from "react";
-
-export class Home extends React.Component {
-    public render() {
-        return <ContentContainerComponent>
-                    <h2>Home</h2>
-                </ContentContainerComponent>;
-    }
-}
