+++ app/core/loading.tsx
@@ -1,0 +1,5 @@
+import * as React from "react";
+
+export function Loading() {
+    return <div>Loading...</div>;
+}
