+++ app/pages/home.tsx
@@ -1,0 +1,8 @@
+import { ContentContainerComponent } from "@justaddjam/strawberry";
+import * as React from "react";
+
+export default function Home() {
+    return <ContentContainerComponent>
+                <h2>Home</h2>
+            </ContentContainerComponent>;
+}
