+++ app/tsconfig.json
@@ -1,7 +1,7 @@
 {
    "compilerOptions": {
       "target": "es5",
-      "module": "es6",
+      "module": "commonjs",
       "moduleResolution": "node",
       "noImplicitAny": true,
       "declaration": false,
